Split the file from src/broker/mod.rs
